fix(modals): rename some flyouts to be shorter
