+++ packages/blockchain-wallet-v4-frontend/src/components/Flyout/RecurringBuy/RemoveConfirm.tsx
