+++ packages/blockchain-wallet-v4-frontend/src/components/Rows/Checkout.tsx
@@ -121,5 +121,4 @@ export type Props = {
   toolTip?: string | React.ReactNode
 }
 
-export { CheckoutRow }
 export default CheckoutRow
