+++ packages/blockchain-wallet-v4-frontend/src/modals/RecurringBuys/RemoveConfirm/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { connect, ConnectedProps } from 'react-redux'
 
-import { RecurringBuyRemoveConfirm } from 'components/Flyout/RecurringBuy'
+import { RemoveConfirm } from 'components/Flyout/RecurringBuy'
 import { actions, selectors } from 'data'
 import { RootState } from 'data/rootReducer'
 import { RecurringBuyRegisteredList, RecurringBuyStepType } from 'data/types'
@@ -11,7 +11,7 @@ import { Props as _P } from '..'
 const Success = ({ activeDetails, close, goBack, removeClick }: Props) => {
   if (!activeDetails) return null
   return (
-    <RecurringBuyRemoveConfirm
+    <RemoveConfirm
       goBack={goBack}
       close={close}
       activeDetails={activeDetails}
